RESTORE_BACKUP_20251014_012206/requirements_monthly_revenue.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0













